moving things into autodapp
