+++ Makefile
@@ -9,16 +9,3 @@ psql:
 mysql:
 	docker exec -it mariadb-server mysql --host=proxy --user=root --password=testpassword testdb
 
-tendermint:
-	docker exec -it tendermint-node /bin/bash
-
-discourse:
-	docker-compose -f demo/discourse/docker-compose.yml up
-
-mediawiki:
-	docker-compose -f demo/mediawiki/docker-compose.yml up
-	
-mediawiki-db:
-	docker exec -it mariadb-server /bin/bash -c "mysql --user=root --password=testpassword --database=testdb < /docker/tables.sql"
-	# docker exec -it mediawiki /bin/bash -c "php /var/www/html/maintenance/importDump.php data.xml"
-
