+++ autodapp/README.md
@@ -1,0 +1,12 @@
+
+# Demos
+
+To run the demos we've implemented (Discourse and Mediawiki),
+you can use the following docker-compose commands:
+
+```bash
+$ make discourse    # discourse demo (based on postgres)
+OR
+$ make mediawiki    # mediawiki demo (based on mariadb)
+$ make mediawiki-db # migrate db
+```
