+++ autodapp/demo/mediawiki/LocalSettings.php
