+++ autodapp/demo/mediawiki/docker-compose.yml
