+++ autodapp/demo/mediawiki/tables.sql
