+++ autodapp/docker-compose.yml
@@ -1,17 +1,6 @@
 version: "3"
 
 services:
-  proxy:
-    image: rust
-    container_name: proxy
-    command: /bin/bash -c "while true; do sleep 200; echo 'TERMINAL is up'; done"
-    volumes:
-      - "./:/code"
-    working_dir: /code
-    links:
-      - mariadb-server
-      - postgres-server
-
   mariadb-server:
     image: mariadb
     container_name: mariadb-server
@@ -44,4 +33,24 @@ services:
       - postgres-server
       - proxy
 
+  proxy:
+    image: rust
+    container_name: proxy
+    command: /bin/bash -c "while true; do sleep 200; echo 'TERMINAL is up'; done"
+    volumes:
+      - "./:/code"
+    working_dir: /code
+    links:
+      - mariadb-server
+      - postgres-server
 
+  tendermint:
+    image: tendermint/tendermint:v0.32.8
+    container_name: tendermint-node
+    entrypoint: /bin/sh -c "tendermint init && tendermint unsafe_reset_all && tendermint node --consensus.create_empty_blocks=false --proxy_app=tcp://proxy:26658 --rpc.laddr=tcp://0.0.0.0:26657"
+    volumes:
+      - "/tmp/tendermint:/tendermint"
+    links:
+      - mariadb-server
+      - postgres-server
+      - proxy
